requirements.txt
aiohttp==3.8.4
aioresponses==0.7.4
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
click==8.1.3
confluent-kafka==2.1.1
ecdsa==0.18.0
fastecdsa==2.3.0
Flask==2.2.3
flask-restx==1.0.6
Flask-SQLAlchemy==3.0.3
freezegun==1.2.2
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.1
pip-autoremove==0.10.0
pluggy==1.0.0
PyMySQL==1.0.3
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.0.2
six==1.16.0
SQLAlchemy==2.0.10
typing_extensions==4.5.0
Werkzeug==2.2.3
yarl==1.9.2
zmq==0.0.0
